requirements.txt
fastapi
uvicorn
prometheus-client
PyYAML
requests
